+++ UniQuanda.Core.Application/CQRS/Commands/Auth/AddExtraEmailForUser/AddExtraEmailForUserDTO.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using UniQuanda.Core.Application.Validators;
 
-namespace UniQuanda.Core.Application.CQRS.Commands.Security.AddExtraEmailForUser;
+namespace UniQuanda.Core.Application.CQRS.Commands.Auth.AddExtraEmailForUser;
 
 public class AddExtraEmailForUserRequestDTO
 {
