+++ UniQuanda.Core.Application/CQRS/Commands/Auth/UpdateUserMainEmail/UpdateUserMainEmailHandler.cs
@@ -3,35 +3,35 @@ using UniQuanda.Core.Application.Repositories;
 using UniQuanda.Core.Application.Services.Auth;
 using UniQuanda.Core.Domain.Enums;
 
-namespace UniQuanda.Core.Application.CQRS.Commands.Security.UpdateUserMainEmail;
+namespace UniQuanda.Core.Application.CQRS.Commands.Auth.UpdateUserMainEmail;
 
 public class UpdateUserMainEmailHandler : IRequestHandler<UpdateUserMainEmailCommand, UpdateResultOfEmailOrPasswordEnum>
 {
-    private readonly ISecurityRepository _securityRepository;
+    private readonly IAuthRepository _authRepository;
     private readonly IPasswordsService _passwordsService;
 
     public UpdateUserMainEmailHandler(
-        ISecurityRepository securityRepository,
+        IAuthRepository authRepository,
         IPasswordsService passwordsService)
     {
-        _securityRepository = securityRepository;
+        _authRepository = authRepository;
         _passwordsService = passwordsService;
     }
 
     public async Task<UpdateResultOfEmailOrPasswordEnum> Handle(UpdateUserMainEmailCommand request, CancellationToken ct)
     {
-        var hashedPassword = await _securityRepository.GetUserHashedPasswordByIdAsync(request.IdUser, ct);
+        var hashedPassword = await _authRepository.GetUserHashedPasswordByIdAsync(request.IdUser, ct);
         if (hashedPassword == null)
             return UpdateResultOfEmailOrPasswordEnum.UserNotExist;
 
         if (!_passwordsService.VerifyPassword(request.PlainPassword, hashedPassword))
             return UpdateResultOfEmailOrPasswordEnum.InvalidPassword;
 
-        var isEmailConnectedWithUser = await _securityRepository.IsEmailConnectedWithUserAsync(request.IdUser, request.NewMainEmail, ct);
+        var isEmailConnectedWithUser = await _authRepository.IsEmailConnectedWithUserAsync(request.IdUser, request.NewMainEmail, ct);
         if (!isEmailConnectedWithUser)
             return UpdateResultOfEmailOrPasswordEnum.EmailNotConnected;
 
-        var updateResult = await _securityRepository.UpdateUserMainEmailAsync(request.IdUser, request.NewMainEmail, ct);
+        var updateResult = await _authRepository.UpdateUserMainEmailAsync(request.IdUser, request.NewMainEmail, ct);
         return updateResult switch
         {
             null => UpdateResultOfEmailOrPasswordEnum.UserNotExist,
