+++ UniQuanda.Core.Application/CQRS/Queries/Auth/GetUserEmails/GetUserEmailsHandler.cs
@@ -1,20 +1,20 @@
 ﻿using MediatR;
 using UniQuanda.Core.Application.Repositories;
 
-namespace UniQuanda.Core.Application.CQRS.Queries.Security.GetUserEmails;
+namespace UniQuanda.Core.Application.CQRS.Queries.Auth.GetUserEmails;
 
 public class GetUserEmailsHandler : IRequestHandler<GetUserEmailsQuery, GetUserEmailsReponseDTO>
 {
-    private readonly ISecurityRepository _securityRepository;
+    private readonly IAuthRepository _authRepository;
 
-    public GetUserEmailsHandler(ISecurityRepository securityRepository)
+    public GetUserEmailsHandler(IAuthRepository authRepository)
     {
-        _securityRepository = securityRepository;
+        _authRepository = authRepository;
     }
 
     public async Task<GetUserEmailsReponseDTO?> Handle(GetUserEmailsQuery request, CancellationToken ct)
     {
-        var userEmails = await _securityRepository.GetUserEmailsAsync(request.IdUser, ct);
+        var userEmails = await _authRepository.GetUserEmailsAsync(request.IdUser, ct);
         if (userEmails is null)
             return null;
