+++ UniQuanda.Core.Application/CQRS/Queries/Auth/GetUserEmails/GetUserEmailsQuery.cs
@@ -1,5 +1,5 @@
 ﻿using MediatR;
-namespace UniQuanda.Core.Application.CQRS.Queries.Security.GetUserEmails;
+namespace UniQuanda.Core.Application.CQRS.Queries.Auth.GetUserEmails;
 
 public class GetUserEmailsQuery : IRequest<GetUserEmailsReponseDTO>
 {
