+++ UniQuanda.Infrastructure/Extensions.cs
@@ -25,7 +25,6 @@ public static class Extensions
         // Repositories
         services.AddScoped<IAuthRepository, AuthRepository>();
         services.AddScoped<IAppUserRepository, AppUserRepository>();
-        services.AddScoped<ISecurityRepository, SecurityRepository>();
 
         // Services
         services.AddScoped<IPasswordsService, PasswordsService>();
